requirements.txt
RPi.GPIO == 0.7.1
Flask == 2.3.3
MIDIUtil == 1.2.1
RPi.GPIO == 0.7.1
midi2audio = 0.1.1
fluidsynth == 0.2

